fix: resolve remaining mypy type errors and fix broken imports
- Fix TYPE_CHECKING imports in 12 pipeline stage run.py files
- Add type:ignore for optional dependencies (wandb, mlflow, catboost, hmmlearn, arch, google.colab, IPython, pydantic, fastapi, uvicorn)
- Fix broken internal imports (manifest, container paths)
- Remove unused type:ignore comments
- Add None checks for config.project_root
- Update JACOB.md with final mypy status
- Apply ruff auto-fixes

Final status: 0 actual code type errors (288 remaining are import-untyped for external libraries)

+++ src/data/pipeline/stages/clean/bar_builders/base.py
@@ -230,7 +230,7 @@ class BaseBarBuilder(ABC):
             compression_ratio=len(df_input) / max(len(df_output), 1),
             min_bars_per_output=min(bars_per_output) if bars_per_output else 0,
             max_bars_per_output=max(bars_per_output) if bars_per_output else 0,
-            avg_bars_per_output=np.mean(bars_per_output) if bars_per_output else 0,
+            avg_bars_per_output=float(np.mean(bars_per_output)) if bars_per_output else 0.0,
         )
 
     def __repr__(self) -> str:
